Samler felles kode i EmbeddedKafkaBase
